requirements/requirements.txt
jax>=0.4.6
jaxtyping>=0.2.14
numpy>=1.22.4
distrax>=0.1.3
dm-haiku>=0.0.9
chex>=0.1.7
